refactor: grv v4 metric class and tests
