+++ ugh3_metrics/metrics/__init__.py
@@ -1,5 +1,7 @@
 from .por_v4 import PorV4, calc_por_v4
 from .deltae_v4 import DeltaEV4, calc_deltae_v4
+from .grv_v4 import GrvV4, calc_grv_v4
 
 __all__ = ["PorV4", "calc_por_v4"]
 __all__.extend(["DeltaEV4", "calc_deltae_v4"])
+__all__.extend(["GrvV4", "calc_grv_v4"])
